restructured validators
